# staffe/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Staffe API
  description: >-
    REST API на Staffe — gestionale di magazzino (WMS) per staffe e accessori
    di fissaggio per ascensori. Estratta dal codice reale in `src/app/api/**`
    (Next.js 15 App Router route handlers) — nessun endpoint qui descritto è
    ipotetico.


    **Formato risposta** (`src/lib/api.ts`, un solo contratto per tutte le
    rotte): successo `{ "data": ..., "meta"?: ... }`; errore
    `{ "error": { "message": "...", "code": "...", "details"?: ... } }`. Il
    messaggio di errore è in italiano ed è mostrabile all'operatore così
    com'è.


    **Autenticazione**: cookie di sessione httpOnly `staffe_session` (JWT
    firmato, `jose`), impostato da `POST /api/auth/login`. Non esiste header
    `Authorization` — il client deve inviare il cookie (`credentials:
    'include'` da browser). Le rotte protette rispondono `401` senza sessione
    valida e `403` se la sessione è valida ma il ruolo non ha il permesso
    richiesto (matrice in `src/lib/rbac.ts`, applicata con
    `requirePermission()`).


    **Paginazione**: le liste accettano `page` (default 1) e `perPage`
    (default variabile per rotta, tetto 200) come query string e rispondono
    con `meta: { page, perPage, total, totalPages }`.


    **Visibilità dei costi**: la risposta dipende dal RUOLO di chi chiama. I
    campi di costo d'acquisto (`unitCostCents` sui movimenti, sulle righe
    d'ordine e sulle righe di ricevimento, più sconti e aliquota d'ordine)
...
servers:
...
security:
  - cookieAuth: []
...
paths:
...
  /api/barcode:
    get:
      tags: [varie]
      summary: Genera un simbolo SVG (code128 / EAN-13 / QR)
      description: Dietro autenticazione (`requireUser`) — non genera in massa senza login.
      parameters:
        - in: query
          name: testo
          required: true
          schema: { type: string, minLength: 1, maxLength: 300 }
        - in: query
          name: tipo
          required: true
          schema: { type: string, enum: [code128, ean13, qr] }
      responses:
        "200":
          description: Simbolo SVG
          content:
            image/svg+xml: { schema: { type: string } }
        "401": { $ref: "#/components/responses/NonAutenticato" }
        "422": { $ref: "#/components/responses/Validazione" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: staffe_session
      description: JWT firmato (HS256, `jose`) in cookie httpOnly + sameSite=lax + secure (produzione).
...
  responses:
...
    NonAutenticato:
      description: Sessione mancante o non valida
      content:
        application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } }
...
    Validazione:
      description: Corpo/parametri non validi (Zod)
      content:
        application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } }
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          required: [message, code]
          properties:
            message: { type: string, description: Messaggio in italiano, mostrabile all'operatore così com'è. }
            code: { type: string, description: "Codice macchina, es. `non_autenticato`, `vietato`, `validazione`, `non_trovato`, `duplicato`, `giacenza`, `stato`." }
            details: {}
...